+++ src/components/Todo/FieldAddTask.vue
@@ -1,0 +1,29 @@
+<template>
+  <v-text-field
+    v-model="newTaskTitle"
+    @click:append="addTask"
+    @keyup.enter="addTask"
+    class="pa-3"
+    outlined
+    label="Add to do"
+    append-icon="mdi-plus"
+    hide-details
+    clearable
+  ></v-text-field>
+</template>
+
+<script>
+export default {
+  data() {
+    return {
+      newTaskTitle: "",
+    };
+  },
+  methods: {
+    addTask() {
+      this.$store.commit("addTask", this.newTaskTitle);
+      this.newTaskTitle = "";
+    },
+  },
+};
+</script>
