+++ src/components/Todo/NoTasks.vue
@@ -1,0 +1,16 @@
+<template>
+  <div class="no-tasks">
+    <v-icon size="100" color="primary darken-2"> mdi-check </v-icon>
+    <div class="primary--text text-h5">No tasks</div>
+  </div>
+</template>
+
+
+<style lang="sass">
+.no-tasks
+    position: absolute
+    left: 50%
+    top: 50%
+    transform: translate(-50%, -50%)
+    opacity: 0.5
+</style>
