+++ src/views/Todo.vue
@@ -1,78 +1,24 @@
 <template>
   <div class="home">
-  <v-text-field
-  v-model="newTaskTitle"
-  @click:append="addTask"
-  @keyup.enter="addTask"
-    class="pa-3"
-    outlined
-    label="Add to do"
-    append-icon="mdi-plus"
-    hide-details
-    clearable
-  ></v-text-field>
 
-    <v-list v-if="$store.state.tasks.length" flat class="pt-0">
-      <div v-for="task in $store.state.tasks" :key="task.id">
-        <v-list-item @click="$store.commit('doneTask', task.id)" :class="{'blue lighten-5': task.done}">
-          <template v-slot:default>
-            <v-list-item-action>
-              <v-checkbox :input-value="task.done"></v-checkbox>
-            </v-list-item-action>
+    <field-add-task />
 
-            <v-list-item-content>
-              <v-list-item-title :class="{'text-decoration-line-through' : task.done}">{{ task.title }}</v-list-item-title>
-            </v-list-item-content>
-            <v-list-item-action>
-          <v-btn icon
-          @click.stop="$store.commit('deleteTask', task.id)"
-          >
-            <v-icon color="primary lighten-1">mdi-delete</v-icon>
-          </v-btn>
-        </v-list-item-action>
-          </template>
-        </v-list-item>
-        <v-divider></v-divider>
-      </div>
-    </v-list>
+    <list-tasks v-if="$store.state.tasks.length" />
+   
+     <no-tasks v-else />
 
-      <div v-else class="no-tasks">
-          <v-icon size= "100" color="primary darken-2">
-            mdi-check
-          </v-icon>
-          <div class="primary--text text-h5">No tasks</div>
-      </div>
   </div>
 </template>
 
 <script>
+
 export default {
   name: "Home",
-  data() {
-    return {
-      newTaskTitle: '',
-      
-    };
-  },
-  methods: {
-    addTask(){      
-
-      this.$store.commit('addTask', this.newTaskTitle)
-      this.newTaskTitle = ''
-
-    },
-    
-    
-  }
+  components: { 'field-add-task': require('@/components/Todo/FieldAddTask.vue').default, 
+                'list-tasks': require('@/components/Todo/ListTasks.vue').default,
+                'no-tasks': require('@/components/Todo/NoTasks.vue').default,
+                },
 }
 </script>
 
-<style lang="sass">
-.no-tasks
-  position: absolute
-  left: 50%
-  top: 50%
-  transform: translate(-50%, -50%)
-  opacity: 0.5
 
-</style>
